+++ src/Audit.NET/AuditEventEnvironment.cs
@@ -1,8 +1,9 @@
-﻿using Newtonsoft.Json;
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace Audit.Core
 {
-    public class AuditEventEnvironment
+    public class AuditEventEnvironment : IAuditOutput
     {
         /// <summary>
         /// Gets or sets the name of the user responsible for the change.
@@ -42,6 +43,26 @@ namespace Audit.Core
         /// <summary>
         /// The locale name
         /// </summary>
+        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
         public string Culture { get; set; }
+
+        [JsonExtensionData]
+        public Dictionary<string, object> CustomFields { get; set; } = new Dictionary<string, object>();
+
+        /// <summary>
+        /// Serializes this Environment entity as a JSON string
+        /// </summary>
+        public string ToJson()
+        {
+            return JsonConvert.SerializeObject(this, Audit.Core.Configuration.JsonSettings);
+        }
+        /// <summary>
+        /// Parses an Environment entity from its JSON string representation.
+        /// </summary>
+        /// <param name="json">JSON string with the Environment entity representation.</param>
+        public static AuditEventEnvironment FromJson(string json)
+        {
+            return JsonConvert.DeserializeObject<AuditEventEnvironment>(json, Audit.Core.Configuration.JsonSettings);
+        }
     }
 }