+++ src/Audit.WebApi/AuditApiAttribute.Core.cs
@@ -18,6 +18,10 @@ namespace Audit.WebApi
         private AuditApiAdapter _adapter = new AuditApiAdapter();
 
         /// <summary>
+        /// Gets or sets a value indicating whether the output should include the Http Response Headers.
+        /// </summary>
+        public bool IncludeResponseHeaders { get; set; }
+        /// <summary>
         /// Gets or sets a value indicating whether the output should include the Http Request Headers.
         /// </summary>
         public bool IncludeHeaders { get; set; }
@@ -75,7 +79,7 @@ namespace Audit.WebApi
             }
             await _adapter.BeforeExecutingAsync(context, IncludeHeaders, IncludeRequestBody, SerializeActionParameters, EventTypeName);
             var actionExecutedContext = await next.Invoke();
-            await _adapter.AfterExecutedAsync(actionExecutedContext, IncludeModelState, ShouldIncludeResponseBody(actionExecutedContext));
+            await _adapter.AfterExecutedAsync(actionExecutedContext, IncludeModelState, ShouldIncludeResponseBody(actionExecutedContext), IncludeResponseHeaders);
         }
 
         private bool ShouldIncludeResponseBody(ActionExecutedContext context)