+++ src/Audit.WebApi/AuditApiAttribute.cs
@@ -16,6 +16,10 @@ namespace Audit.WebApi
         private AuditApiAdapter _adapter = new AuditApiAdapter();
 
         /// <summary>
+        /// Gets or sets a value indicating whether the output should include the Http Response Headers.
+        /// </summary>
+        public bool IncludeResponseHeaders { get; set; }
+        /// <summary>
         /// Gets or sets a value indicating whether the output should include the Http Request Headers.
         /// </summary>
         public bool IncludeHeaders { get; set; }
@@ -92,7 +96,7 @@ namespace Audit.WebApi
             {
                 return;
             }
-            await _adapter.AfterExecutedAsync(actionExecutedContext, GetContextWrapper(actionExecutedContext.Request), IncludeModelState, ShouldIncludeResponseBody(actionExecutedContext));
+            await _adapter.AfterExecutedAsync(actionExecutedContext, GetContextWrapper(actionExecutedContext.Request), IncludeModelState, ShouldIncludeResponseBody(actionExecutedContext), IncludeResponseHeaders);
         }
 
         private bool ShouldIncludeResponseBody(HttpActionExecutedContext context)