+++ src/Audit.WebApi/ConfigurationApi/IAuditMiddlewareConfigurator.cs
@@ -12,14 +12,24 @@ namespace Audit.WebApi.ConfigurationApi
         /// <param name="requestPredicate">Return true to include the request on the audit output, or false otherwise.</param>
         IAuditMiddlewareConfigurator FilterByRequest(Func<HttpRequest, bool> requestPredicate);
         /// <summary>
-        /// Specifies whether the HTTP headers should be included on the audit output.
+        /// Specifies whether the HTTP Response headers should be included on the audit output.
         /// </summary>
-        /// <param name="include">True to include the HTTP headers, false otherwise</param>
+        /// <param name="include">True to include the HTTP Response headers, false otherwise</param>
+        IAuditMiddlewareConfigurator IncludeResponseHeaders(bool include = true);
+        /// <summary>
+        /// Specifies a predicate to determine whether the HTTP Response headers should be included on the audit output.
+        /// </summary>
+        /// <param name="includePredicate">A function of the executing context to determine whether the HTTP Response headers should be included on the audit output</param>
+        IAuditMiddlewareConfigurator IncludeResponseHeaders(Func<HttpContext, bool> includePredicate);
+        /// <summary>
+        /// Specifies whether the HTTP Request headers should be included on the audit output.
+        /// </summary>
+        /// <param name="include">True to include the HTTP Request headers, false otherwise</param>
         IAuditMiddlewareConfigurator IncludeHeaders(bool include = true);
         /// <summary>
-        /// Specifies a predicate to determine whether the HTTP headers should be included on the audit output.
+        /// Specifies a predicate to determine whether the HTTP Request headers should be included on the audit output.
         /// </summary>
-        /// <param name="includePredicate">A function of the executing context to determine whether the HTTP headers should be included on the audit output</param>
+        /// <param name="includePredicate">A function of the executing context to determine whether the HTTP Request headers should be included on the audit output</param>
         IAuditMiddlewareConfigurator IncludeHeaders(Func<HttpContext, bool> includePredicate);
         /// <summary>
         /// Specifies whether the request body should be included on the audit output.