+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [14.0.2] - 2018-12-15
+### Added
+- Audit.WebApi and Audit.WebApi.Core: Added Response Headers to the event output (optional by `IncludeResponseHeaders` configuration, default is _false_). 
+- Audit.Core: Added optional Custom Fields to `AuditEvent.Environment`.
+
 ## [14.0.1] - 2018-11-28
 ### Added
 - Audit.EntityFramework and Audit.EntityFramework.Core: Added Schema property on entries to complement table name (#182).